cloud_reid_functions/requirements.txt
flask
google-cloud-pubsub
google-cloud-datastore
pydantic